weapon-warehouse-system/app/requirements.txt
fastapi
pandas
python-multipart
uvicorn
pydantic-settings
mysql-connector-python
pydantic